+++ client/src/lib/json-to-iac/constants.js
@@ -12,7 +12,7 @@ terraform {
   required_providers {
     ibm = {
       source  = "IBM-Cloud/ibm"
-      version = "~>1.61.0"
+      version = "~>1.63.0"
     }
 $ADDITIONAL_PROVIDERS
   }