+++ deploy/power_vs_workspaces/versions.tf
@@ -6,7 +6,7 @@ terraform {
   required_providers {
     ibm = {
       source  = "IBM-Cloud/ibm"
-      version = "~>1.61.0"
+      version = "~>1.63.0"
     }
   }
   required_version = ">=1.4"